Refactor: Component logic to hooks
