+++ frontend/src/components/Pages/YTC/index.tsx
@@ -1,7 +1,4 @@
-import React, { useEffect, useState } from "react"
-import { getBaseTokensWithActiveTranches } from "../../../features/element";
-import { elementAddressesAtom } from "../../../recoil/element/atom";
-import { Token } from "../../../types/manual/types";
+import React from "react"
 import { Calculator } from "./Calculator";
 import { Execution } from "./Executor";
 import { useRecoilValue } from 'recoil';
@@ -17,16 +14,8 @@ interface YTCProps {}
 
 export const YTC: React.FC<YTCProps> = (props) => {
 
-    const [baseTokens, setBaseTokens] = useState<Token[]>([]);
     const resultIndex = useRecoilValue(selectedSimulationAtom);
     const simulationResults: YTCOutput[] = useRecoilValue(calculatorGainSelector);
-    const elementAddresses = useRecoilValue(elementAddressesAtom);
-
-    useEffect(() => {
-        getBaseTokensWithActiveTranches(elementAddresses).then((res) => {
-            setBaseTokens(res);
-        })
-    }, [elementAddresses])
 
     return <div>
         <Title
@@ -35,7 +24,6 @@ export const YTC: React.FC<YTCProps> = (props) => {
             infoLink="https://medium.com/@component_general/how-to-yield-token-compound-using-the-ytc-tool-742d140a7c9c"
         />
                     <Calculator
-                        tokens={baseTokens}
                     />
                     {
                         (simulationResults.length > 0) && <>
