+++ frontend/src/components/Prices/index.tsx
@@ -1,12 +1,7 @@
-import React, { useEffect, useState } from 'react'
-import { elementAddressesAtom } from '../../recoil/element/atom';
-import {useRecoilValue} from 'recoil';
-import { getTokenPrice } from '../../features/prices';
+import React from 'react'
 import { shortenNumber } from '../../utils/shortenNumber';
-import { getYTCSpotPrice } from '../../features/element/ytcSpot';
 import { Text } from '@chakra-ui/react';
-import { Web3Provider } from '@ethersproject/providers';
-import { useWeb3React } from '@web3-react/core';
+import { useBaseTokenPrice, useYieldTokenPrice } from './hooks';
 
 interface PriceFeedProps {
     price: number | undefined;
@@ -32,7 +27,6 @@ const PriceTag: React.FC<PriceFeedProps> = (props) => {
         value = "$0"
     }
 
-
     return (
         <>
             {value}
@@ -47,22 +41,10 @@ interface YTPriceTagProps {
 }
 
 export const YTPriceTag: React.FC<YTPriceTagProps> = (props) => {
-    const {baseTokenName, trancheAddress, amount} = props;
 
-    const [price, setPrice] = useState<number>(0);
-    const { library } = useWeb3React();
-    const provider = library as Web3Provider;
-    const elementAddresses = useRecoilValue(elementAddressesAtom);
+    const {amount, baseTokenName, trancheAddress} = props;
 
-    useEffect(() => {
-        const signer = provider?.getSigner();
-        if (baseTokenName && trancheAddress && signer){
-            getYTCSpotPrice(baseTokenName, trancheAddress, elementAddresses, signer).then((price) => {
-                setPrice(price);
-            })
-        }
-    }, [provider, baseTokenName, trancheAddress, elementAddresses])
-
+    const {price} = useYieldTokenPrice(baseTokenName, trancheAddress);
 
     return (
         <BaseTokenPriceTag
@@ -78,27 +60,8 @@ interface BaseTokenPriceTagProps {
 }
 
 export const BaseTokenPriceTag: React.FC<BaseTokenPriceTagProps> = (props) => {
-    const {amount, baseTokenName} = props;
-    const [isLoading, setIsLoading] = useState<boolean>(false);
-
-    const [price, setPrice] = useState<number>(0);
-    const { library } = useWeb3React();
-    const provider = library as Web3Provider;
-    const elementAddresses = useRecoilValue(elementAddressesAtom);
-
-    useEffect(() => {
-        const signer = provider?.getSigner();
-        if(baseTokenName && signer){
-            setIsLoading(true);
-            getTokenPrice(baseTokenName, elementAddresses, signer).then((value) => {
-                setPrice(value);
-            }).catch((error) => {
-                console.error(error);
-            }).finally(() => {
-                setIsLoading(false)
-            })
-        }
-    }, [baseTokenName, elementAddresses, provider])
+    const {amount, baseTokenName} = props
+    const {price, isLoading} = useBaseTokenPrice(baseTokenName)
 
 
     return (
