+++ frontend/src/features/element/fixedRate.ts
@@ -1,6 +1,6 @@
 // calculate the fixed rate of the pToken
 
-import { Signer } from "ethers";
+import { ethers, Signer } from "ethers";
 import { ElementAddresses, Tranche } from "../../types/manual/types";
 import { getReserves } from "../../utils/element/getReserves";
 import { calcSpotPricePt } from "../../utils/element/calcSpotPrice";
@@ -10,7 +10,7 @@ import { ONE_YEAR_IN_SECONDS } from "../../constants/time";
 
 // calculate the fixed rate of the pTokens for a tranche
 
-export const getFixedRate = async (tokenName: string, trancheAddress: string, elementAddresses: ElementAddresses, signer: Signer): Promise<number> => {
+export const getFixedRate = async (tokenName: string, trancheAddress: string, elementAddresses: ElementAddresses, signerOrProvider: Signer | ethers.providers.Provider): Promise<number> => {
     const tranche = _.find(elementAddresses.tranches[tokenName], (tranche: Tranche) => (tranche.address === trancheAddress));
     const balancerAddress = elementAddresses.balancerVault;
 
@@ -22,7 +22,7 @@ export const getFixedRate = async (tokenName: string, trancheAddress: string, el
 
     const ptPool = tranche.ptPool;
 
-    const reserves = await getReserves(ptPool.address, balancerAddress, signer);
+    const reserves = await getReserves(ptPool.address, balancerAddress, signerOrProvider);
 
     const tParamSeconds = ptPool.timeStretch * ONE_YEAR_IN_SECONDS;
