+++ frontend/src/features/element/principalTotal.ts
@@ -6,9 +6,9 @@ import ERC20 from '../../artifacts/contracts/balancer-core-v2/lib/openzeppelin/E
 import { ERC20 as ERC20Type} from '../../hardhat/typechain/ERC20';
 
 
-export const getPrincipalTotal = async (elementAddresses: ElementAddresses, trancheAddress: string, signer: Signer): Promise<number> => {
-    const tranche = new Contract(trancheAddress, ITranche.abi, signer) as ITrancheType;
-    const trancheERC20 = new Contract(trancheAddress, ERC20.abi, signer) as ERC20Type;
+export const getPrincipalTotal = async (elementAddresses: ElementAddresses, trancheAddress: string, signerOrProvider: Signer | ethers.providers.Provider): Promise<number> => {
+    const tranche = new Contract(trancheAddress, ITranche.abi, signerOrProvider) as ITrancheType;
+    const trancheERC20 = new Contract(trancheAddress, ERC20.abi, signerOrProvider) as ERC20Type;
 
     const supply = await tranche.totalSupply();
