+++ frontend/src/features/element/wrappedPositionAmount.ts
@@ -9,14 +9,14 @@ import { ElementAddresses } from '../../types/manual/types';
 import ERC20 from '../../artifacts/contracts/balancer-core-v2/lib/openzeppelin/ERC20.sol/ERC20.json';
 import { ERC20 as ERC20Type} from '../../hardhat/typechain/ERC20';
 
-export const getUnderlyingTotal = async (elementAddresses: ElementAddresses, trancheAddress: string, signer: Signer): Promise<number> => {
+export const getUnderlyingTotal = async (elementAddresses: ElementAddresses, trancheAddress: string, signerOrProvider: Signer | ethers.providers.Provider): Promise<number> => {
     // get the tranche
-    const tranche = new Contract(trancheAddress, ITranche.abi, signer) as ITrancheType;
+    const tranche = new Contract(trancheAddress, ITranche.abi, signerOrProvider) as ITrancheType;
 
     // get the wrapped position address
     const wrappedPositionAddress = await tranche.position();
-    const wrappedPosition = new Contract(wrappedPositionAddress, IWrappedPosition.abi, signer) as IWrappedPositionType;
-    const wrappedPositionERC20 = new Contract(wrappedPositionAddress, ERC20.abi, signer) as ERC20Type;
+    const wrappedPosition = new Contract(wrappedPositionAddress, IWrappedPosition.abi, signerOrProvider) as IWrappedPositionType;
+    const wrappedPositionERC20 = new Contract(wrappedPositionAddress, ERC20.abi, signerOrProvider) as ERC20Type;
 
     // get the balance of the tranche for the wrapped position
     const balance = await wrappedPosition.balanceOfUnderlying(trancheAddress);
