+++ frontend/src/features/element/ytcSpot.ts
@@ -1,10 +1,10 @@
-import { Signer } from "ethers";
+import { ethers, Signer } from "ethers";
 import _ from "lodash";
 import { ElementAddresses, Tranche } from "../../types/manual/types";
 import { calcSpotPriceYt } from "../../utils/element/calcSpotPrice";
 import { getReserves } from "../../utils/element/getReserves";
 
-export const getYTCSpotPrice = async (tokenName: string, trancheAddress: string, elementAddresses: ElementAddresses, signer: Signer): Promise<number> => {
+export const getYTCSpotPrice = async (tokenName: string, trancheAddress: string, elementAddresses: ElementAddresses, signerOrProvider: Signer | ethers.providers.Provider): Promise<number> => {
     const tranche = _.find(elementAddresses.tranches[tokenName], (tranche: Tranche) => (tranche.address === trancheAddress));
     const balancerAddress = elementAddresses.balancerVault;
 
@@ -16,7 +16,7 @@ export const getYTCSpotPrice = async (tokenName: string, trancheAddress: string,
 
     const ytPool = tranche.ytPool;
 
-    const reserves = await getReserves(ytPool.address, balancerAddress, signer);
+    const reserves = await getReserves(ytPool.address, balancerAddress, signerOrProvider);
 
     // The getReserves function does not return the base/yt reserves in a particular order
     // It does return the token address in tokens in the same index as it's balance in "balance"
