+++ frontend/src/recoil/element/atom.ts
@@ -1,5 +1,7 @@
-import {atom} from 'recoil'
+import {atom, selector} from 'recoil'
 import constants from '../../hardhat/mainnet-constants.json';
+import { getBaseTokensWithActiveTranches } from "../../features/element";
+import { Token } from '../../types/manual/types';
 
 // This is a recoil atom
 // Atoms are pieces of state that can be accessed and or modified by various components through a set of hooks
@@ -9,4 +11,15 @@ import constants from '../../hardhat/mainnet-constants.json';
 export const elementAddressesAtom = atom({
   key: 'elementAddresses',
   default: constants,
+})
+
+export const activeTokensSelector = selector<Token[]>({
+  key: 'activeTokens',
+  get: ({get}) => {
+    const elementAddresses = get(elementAddressesAtom);
+    
+    const activeTokens =  getBaseTokensWithActiveTranches(elementAddresses);
+
+    return activeTokens;
+  }
 })
